fastsaliency_toolbox/requirements.txt
torch #==1.9.0
torchvision #==0.10.0
numpy==1.21.2
pillow==8.3.2
scipy==1.7.1
click
scikit-image==0.18.3
streamlit==0.89.0
scikit-learn
# Successfully installed PyWavelets-1.1.1 cycler-0.10.0 imageio-2.9.0 kiwisolver-1.3.2 matplotlib-3.4.3 networkx-2.6.3 pyparsing-2.4.7 python-dateutil-2.8.2 scikit-image-0.18.3 six-1.16.0 tifffile-2021.8.30
